[node-states-viz] allow selected genotype to be a (spatial) resolution
Previously it didn't make sense for a selected genotype to be a geo-resolution as there's no logical mapping between a nucleotide/aa and a lat-long. However for a force-directed state-map we do want to be able to visualise these.

This changes the behavior so that when a genotype is selected a new option appears in the available geo-resolutions. If the color-by changes to a non-genotype, then this option is removed. If the selected geo-res is the genotype, and the color-by changes, we either update the geo res to the new genotype color-by or switch to the default geo res.

This necessitated refactoring `getTraitFromNode` to handle genotype traits, and allowed us to get rid of the `getTipColorAttribute` function.